Added Hitbox, Hurtbox, and attacking
Added a Hitbox for attacking and Hurtbox for being attacked, both derived from Area2D. Added a player attack animation and hooked up hitbox to it. Made a diamond in test scene with a hurtbox to attack.

+++ Scripts/Combat/IHealth.cs
@@ -3,6 +3,6 @@ public interface IHealth
     int MaxHealth { get; }
     float CurrHealth { get; }
 
-    void TakeDamage(int amount);
-    void Heal(int amount);
+    void TakeDamage(float amount);
+    void Heal(float amount);
 }
